Title Screen and Options Menu
I created a basic version of our title screen and option menu. The title screen allows you to go to the options menu, start the game and quit the game. The options menu only lets you go back to the title screen and does not let you customise anything yet.